phase_1/backend/requirements.txt
Flask
psycopg2-binary
gunicorn
pandas
numpy
SQLAlchemy
Flask-SQLAlchemy
python-dotenv
python-dotenv